requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
idna==2.8
mysqlclient==1.4.6
newsapi-python==0.2.6
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.3
xlwt==1.3.0
